rust: rename log.rs to sclog.rs
To avoid conflict with the Rust log crate.
