+++ rust/src/applayertemplate/template.rs
@@ -17,7 +17,7 @@
 
 use std;
 use crate::core::{self, ALPROTO_UNKNOWN, AppProto, Flow, IPPROTO_TCP};
-use crate::log::*;
+use crate::sclog::*;
 use std::mem::transmute;
 use crate::applayer::{self, LoggerFlags};
 use crate::parser::*;
