+++ rust/src/dns/dns.rs
@@ -20,7 +20,7 @@ extern crate nom;
 use std;
 use std::mem::transmute;
 
-use crate::log::*;
+use crate::sclog::*;
 use crate::applayer::LoggerFlags;
 use crate::core;
 use crate::dns::parser;
