+++ rust/src/ftp/mod.rs
@@ -23,7 +23,7 @@ use std::str;
 use std;
 use std::str::FromStr;
 
-use crate::log::*;
+use crate::sclog::*;
 
 // We transform an integer string into a i64, ignoring surrounding whitespaces
 // We look for a digit suite, and try to convert it.
