+++ rust/src/nfs/nfs2.rs
@@ -18,7 +18,7 @@
 // written by Victor Julien
 
 use nom;
-use crate::log::*;
+use crate::sclog::*;
 
 use crate::nfs::nfs::*;
 use crate::nfs::types::*;
