+++ rust/src/nfs/nfs3.rs
@@ -18,7 +18,7 @@
 // written by Victor Julien
 
 use nom;
-use crate::log::*;
+use crate::sclog::*;
 use crate::core::*;
 
 use crate::nfs::nfs::*;
