+++ rust/src/smb/smb1_session.rs
@@ -15,7 +15,7 @@
  * 02110-1301, USA.
  */
 
-use crate::log::*;
+use crate::sclog::*;
 
 use crate::smb::smb_records::*;
 use crate::smb::smb1_records::*;
