+++ rust/src/smb/smb2.rs
@@ -18,7 +18,7 @@
 use nom;
 
 use crate::core::*;
-use crate::log::*;
+use crate::sclog::*;
 
 use crate::smb::smb::*;
 use crate::smb::smb2_records::*;
